// components/profile/types.ts
export interface ProfileForm {
  firstName: string;
  lastName: string;
  phone: string;

  headline: string;
  bio: string;

  address: string;
  city: string;
  state: string;
  country: string;
  pincode: string;

  college: string;
  university: string;
  degree: string;
  branch: string;
  passingYear: string;

  currentCompany: string;
  designation: string;
  experience: string;

  linkedin: string;
  github: string;
  portfolio: string;
  website: string;
}